refactor(env): centralize URLs in environment variables for portal redirects
- Added NEXT_PUBLIC_USER_APP_URL to admin app environment example
- Updated imports in next-env.d.ts files for correct routes typing
- Added id attribute to main content container in admin layout
- Introduced request timeout and response validation in admin login mutation
- Replaced hardcoded URLs with environment variables for user and admin portals
- Updated client-side redirects to use environment URLs or fallback defaults
- Modified user dashboard to conditionally enable queries based on authentication
- Improved login page error messages with dynamic admin and user portal URLs
- Refactored auth hooks to redirect users using environment-based URLs instead of hardcoded ones
