+++ apps/admin/.env.example
@@ -1,1 +1,2 @@
 NEXT_PUBLIC_API_URL=http://localhost:3333/api/v1
+NEXT_PUBLIC_USER_APP_URL=http://localhost:3000
