+++ apps/admin/src/app/layout.tsx
@@ -25,7 +25,10 @@ export default function RootLayout({
           <ChakraUIProvider>
             <div className="flex min-h-screen bg-gray-900">
               <AdminSidebar />
-              <main className="flex-1 md:ml-64 p-4 md:p-8 transition-all duration-200">
+              <main
+                id="main-content"
+                className="flex-1 md:ml-64 p-4 md:p-8 transition-all duration-200"
+              >
                 {children}
               </main>
             </div>
