+++ apps/user/src/app/login/metadata.ts
@@ -1,9 +1,7 @@
-import { Metadata } from 'next';
+import type { Metadata } from 'next';
 
 export const metadata: Metadata = {
   title: 'Login | IELTS Master',
   description:
     'Sign in to your IELTS Master account to continue your vocabulary learning journey.',
 };
-
-export { default } from './page';
